requirements.txt
tensorflow>=2.12
ultralytics>=8.0.0
cryptography==45.0.3
opencv-python
matplotlib
numpy
pillow
tifffile
scikit-image
psutil
ipython
flask
PyJWT